# Remove one unnecessary copy of the output during the type promotion.